# src/utility/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music API",
    "description": "API per il sito Social Network for Music",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/login": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login di un utente già registrato (genera un token JWT valido per 12 ore ) ",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/registrazione": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registrazione di un nuovo utente",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "cognome": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/detete-account": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Elimina l'account dell'utente",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-user-info": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Restituisce le informazioni dell'utente",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add-favourite-artist": {
      "post": {
        "tags": [
          "User favourite"
        ],
        "summary": "Aggiunge un artista ai preferiti dell'utente",
        "description": "",
        "parameters": [
          {
            "name": "artistId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-favourite-artists": {
      "get": {
        "tags": [
          "User favourite"
        ],
        "summary": "Restituisce gli artisti preferiti dell'utente",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/remove-favourite-artist": {
      "delete": {
        "tags": [
          "User favourite"
        ],
        "summary": "Rimuove un artista dai preferiti dell'utente",
        "description": "",
        "parameters": [
          {
            "name": "artistId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add-favourite-genre": {
      "post": {
        "tags": [
          "User favourite"
        ],
        "summary": "Aggiunge un genere ai preferiti dell'utente",
        "description": "",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-favourite-genres": {
      "get": {
        "tags": [
          "User favourite"
        ],
        "summary": "Restituisce i generi preferiti dell'utente",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/remove-favourite-genre": {
      "delete": {
        "tags": [
          "User favourite"
        ],
        "summary": "Rimuove un genere dai preferiti dell'utente",
        "description": "",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/update-user-info": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Aggiorna le informazioni dell'utente",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "cognome": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}